#63 - refactoring: new package with config/settings files
