+++ src/main/java/de/bayern/gdi/App.java
@@ -18,7 +18,7 @@
 package de.bayern.gdi;
 
 import de.bayern.gdi.gui.Start;
-import de.bayern.gdi.utils.Config;
+import de.bayern.gdi.config.Config;
 
 import java.io.IOException;
 import java.lang.management.ManagementFactory;
