+++ src/main/java/de/bayern/gdi/config/ApplicationSettings.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package de.bayern.gdi.utils;
+package de.bayern.gdi.config;
 
 import java.io.IOException;
 import java.util.HashMap;
