+++ src/main/java/de/bayern/gdi/gui/DataBean.java
@@ -26,9 +26,9 @@ import de.bayern.gdi.services.CatalogService;
 import de.bayern.gdi.services.Service;
 import de.bayern.gdi.services.ServiceType;
 import de.bayern.gdi.services.WFSMeta;
-import de.bayern.gdi.utils.Config;
+import de.bayern.gdi.config.Config;
 import de.bayern.gdi.utils.ServiceChecker;
-import de.bayern.gdi.utils.ServiceSettings;
+import de.bayern.gdi.config.ServiceSettings;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
