+++ src/main/java/de/bayern/gdi/gui/controller/Controller.java
@@ -40,11 +40,11 @@ import de.bayern.gdi.processor.ProcessorListener;
 import de.bayern.gdi.services.FilterEncoder;
 import de.bayern.gdi.services.ServiceType;
 import de.bayern.gdi.services.WFSMeta;
-import de.bayern.gdi.utils.Config;
+import de.bayern.gdi.config.Config;
 import de.bayern.gdi.utils.DownloadConfig;
 import de.bayern.gdi.utils.I18n;
 import de.bayern.gdi.utils.ServiceChecker;
-import de.bayern.gdi.utils.ServiceSettings;
+import de.bayern.gdi.config.ServiceSettings;
 import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
