+++ src/main/java/de/bayern/gdi/gui/controller/FilterAtomController.java
@@ -30,7 +30,7 @@ import de.bayern.gdi.gui.map.PolygonClickedEvent;
 import de.bayern.gdi.gui.map.PolygonInfos;
 import de.bayern.gdi.services.Atom;
 import de.bayern.gdi.utils.I18n;
-import de.bayern.gdi.utils.ServiceSettings;
+import de.bayern.gdi.config.ServiceSettings;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
