+++ src/main/java/de/bayern/gdi/gui/controller/MenuBarController.java
@@ -18,7 +18,7 @@
 package de.bayern.gdi.gui.controller;
 
 import de.bayern.gdi.gui.WebViewWindow;
-import de.bayern.gdi.utils.Config;
+import de.bayern.gdi.config.Config;
 import de.bayern.gdi.utils.DownloadConfig;
 import de.bayern.gdi.utils.I18n;
 import de.bayern.gdi.utils.Misc;
