+++ src/main/java/de/bayern/gdi/gui/controller/ProcessingChainController.java
@@ -21,7 +21,7 @@ import de.bayern.gdi.gui.UIFactory;
 import de.bayern.gdi.model.MIMEType;
 import de.bayern.gdi.model.MIMETypes;
 import de.bayern.gdi.model.ProcessingStepConfiguration;
-import de.bayern.gdi.utils.Config;
+import de.bayern.gdi.config.Config;
 import de.bayern.gdi.utils.DownloadConfig;
 import de.bayern.gdi.utils.I18n;
 import javafx.collections.ObservableList;
