+++ src/main/java/de/bayern/gdi/gui/map/MapHandler.java
@@ -26,9 +26,9 @@ import com.sothawo.mapjfx.MapType;
 import com.sothawo.mapjfx.MapView;
 import com.sothawo.mapjfx.WMSParam;
 import com.sothawo.mapjfx.event.MapViewEvent;
-import de.bayern.gdi.utils.Config;
+import de.bayern.gdi.config.Config;
 import de.bayern.gdi.utils.I18n;
-import de.bayern.gdi.utils.ServiceSettings;
+import de.bayern.gdi.config.ServiceSettings;
 import javafx.application.Platform;
 import javafx.event.EventTarget;
 import javafx.scene.control.Button;
