+++ src/test/java/de/bayern/gdi/utils/AutoFileNamesTest.java
@@ -17,6 +17,7 @@
  */
 package de.bayern.gdi.utils;
 
+import de.bayern.gdi.config.Config;
 import junit.framework.TestCase;
 import org.junit.Before;
 import org.junit.Test;
